test(deno bundle): add sanity check for deno bundle for v1.7.1+
The just-released Deno 1.7.1 appears to resolve the issues
that were causing invalid bundles to be generated for Zod.
This PR adds a sanity check that the generated bundle is at
least importable without an error, running against Deno 1.7.1
and the latest stable release.

Also adds yarn clean to fix.yaml to remove stale files.